[+] Add color slider

+++ src/wasm/constants/mod.rs
@@ -1,3 +1,3 @@
 pub const MAX_H: f32 = 360.0;
-pub const MAX_S: f32 = 100.0;
-pub const MAX_V: f32 = 100.0;
+pub const MAX_S: i32 = 100;
+pub const MAX_V: i32 = 100;
